# ecommerce-api/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:9292/v1/api/ecommerce-api","description":"Generated server url"}],"tags":[{"name":"Productos","description":"API para gestión de productos"},{"name":"Pedidos","description":"API para gestión de pedidos"}],"paths":{"/producto/{id}":{"put":{"tags":["Productos"],"summary":"Actualizar producto existente","operationId":"actualizarProducto","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"description":"Datos para actualizar el producto","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductoResponse"}}}}}}},"/producto":{"get":{"tags":["Productos"],"summary":"Listar todos los productos","operationId":"listarProductos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductoResponse"}}}}}}},"post":{"tags":["Productos"],"summary":"Crear nuevo producto","operationId":"crearProducto","requestBody":{"description":"Datos necesarios para crear un producto","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductoRequest"}}},"required":true},"responses":{"201":{"description":"Producto creado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductoResponse"}}}}}}},"/pedido":{"get":{"tags":["Pedidos"],"summary":"Listar todos los pedidos","operationId":"listarPedidos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PedidoResponse"}}}}}}},"post":{"tags":["Pedidos"],"summary":"Crear nuevo pedido","operationId":"crearPedido","requestBody":{"description":"Lista de productos que formarán el pedido","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PedidoRequest"}}},"required":true},"responses":{"201":{"description":"Pedido creado correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PedidoResponse"}}}}}}},"/pedido/{id}/confirmar":{"post":{"tags":["Pedidos"],"summary":"Confirmar pedido y aplicar descuentos","operationId":"confirmarPedido","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PedidoResponse"}}}}}}}},"components":{"schemas":{"ProductoRequest":{"type":"object","description":"Datos necesarios para crear o actualizar un producto","properties":{"nombre":{"type":"string","description":"Nombre del producto","example":"Laptop Gamer"},"precio":{"type":"number","description":"Precio del producto","example":2500.0},"stock":{"type":"integer","format":"int32","description":"Cantidad disponible en stock","example":10}}},"ProductoResponse":{"type":"object","description":"Información devuelta de un producto","properties":{"id":{"type":"integer","format":"int64","description":"ID único del producto","example":1},"nombre":{"type":"string","description":"Nombre del producto","example":"Laptop Gamer"},"precio":{"type":"number","description":"Precio actual del producto","example":2500.0},"stock":{"type":"integer","format":"int32","description":"Cantidad disponible en stock","example":10}}},"PedidoItemRequest":{"type":"object","description":"Producto incluido en el pedido","properties":{"productoId":{"type":"integer","format":"int64","description":"ID del producto","example":1},"cantidad":{"type":"integer","format":"int32","description":"Cantidad solicitada","example":2}}},"PedidoRequest":{"type":"object","description":"Pedido con lista de productos","properties":{"items":{"type":"array","description":"Lista de productos incluidos en el pedido","items":{"$ref":"#/components/schemas/PedidoItemRequest"}}}},"PedidoItemResponse":{"type":"object","description":"Detalle de un producto dentro del pedido","properties":{"productoId":{"type":"integer","format":"int64","description":"ID del producto","example":1},"cantidad":{"type":"integer","format":"int32","description":"Cantidad solicitada","example":2},"precioUnitario":{"type":"number","description":"Precio unitario del producto","example":2500.0},"subtotal":{"type":"number","description":"Subtotal del item (precio * cantidad)","example":5000.0}}},"PedidoResponse":{"type":"object","description":"Información completa de un pedido","properties":{"id":{"type":"integer","format":"int64","description":"ID único del pedido","example":100},"fecha":{"type":"string","format":"date-time","description":"Fecha de creación del pedido","example":"2026-02-14T15:30:00"},"items":{"type":"array","description":"Lista de productos incluidos en el pedido","items":{"$ref":"#/components/schemas/PedidoItemResponse"}},"total":{"type":"number","description":"Total bruto sin descuentos","example":5000.0},"estado":{"type":"string","description":"Estado actual del pedido","example":"CONFIRMADO"},"descuento":{"type":"number","description":"Descuentos","example":500.0}}}}}}